📎 :: AuthActivity
