+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.Helper"
         tools:targetApi="31" >
         <activity
-            android:name=".splash.SplashActivity"
+            android:name=".AuthActivity"
             android:exported="true"
             android:theme="@style/Theme.Helper">
             <intent-filter>
